Changed Some Comment
Overall the code looked good.  It was easy to follow and understand.
You used good variable names.  I just added a comment to your function
that generates your randomn shapes.  I also moved your comments with
your code.  This made it easier to follow and I was not having to go
back and forth.